backend/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.27,<1
pydantic>=2.5,<3
pydantic-settings>=2.1,<3
python-jose[cryptography]>=3.3,<4
httpx>=0.26,<1
supabase>=2.3,<3
anthropic>=0.40,<1
google-generativeai>=0.8,<1
langchain>=0.3,<1
langchain-anthropic>=0.3,<1
pgvector>=0.3,<1
python-multipart>=0.0.7
python-dotenv>=1.0,<2
